Refactor wizards file structure (#130)
