+++ src/wizards/tokenSwapDealWizard/components/proposalLeadSection/proposalLeadSection.ts
@@ -1,6 +1,6 @@
 import { bindable } from "aurelia-typed-observable-plugin";
 import { IProposalLead } from "entities/DealRegistrationTokenSwap";
-import { WizardErrors } from "services/WizardService";
+import { WizardErrors } from "wizards/services/WizardService";
 
 export class ProposalLeadSection {
   @bindable errors: WizardErrors<IProposalLead>;
