+++ src/wizards/tokenSwapDealWizard/components/stageWrapper/stageWrapper.scss
